Initial commit: "version" and "authenticate" messages
This code is already pretty messy but I found a bug in Mumble 1.3.0
with it, so I wanted to commit it and push it somewhere so that I could
refer back to it if I needed to.

Right now I only have the client-side of the protocol, and it only
supports the "version" and "authenticate" messages (over TLS). The
client will connect to a standard Mumble server, but it doesn't ping so
the server will disconnect you after 30 seconds.

The client doesn't implement any other part of the protocol yet, and
doesn't have any audio capabilities.

Also, the TLS code pretty much doesn't check anything about the
encryption, so it's pretty insecure. This will be fixed later, though.