Networking with non-blocking TCP streams.
I think the non-blocking part is non-ideal, since it required creating
an Arc<Mutex<T>> to share the stream. I'm going to try cloning the
TcpStream and feeding those to two new TlsStreams so they can both
read/write independently, and just use one for reading and one for
writing. Is that how networking works? idontknowwhatimdoing.jpg

This version also pings every once in a while to stay alive in the
server.